Further clean up 'make test'
Use the same input file for both C and C++.
Verify that `cdecl` gives warnings on C++ constructs.
Add fflush in cdecl.c so that the warnings on stderr will come out
properly interleaved with the ordinary output on stdout, for the
benefit of the tests.
Fix up ALLFILES (which had gotten out of sync), not that I think it's
really necessary anymore anyway.